+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/IssuedCode.kt
@@ -1,0 +1,12 @@
+package at.asitplus.wallet.lib.oauth2
+
+import at.asitplus.openid.AuthorizationDetails
+import at.asitplus.openid.OidcUserInfoExtended
+
+data class IssuedCode(
+    val code: String,
+    val userInfoExtended: OidcUserInfoExtended,
+    val scope: String? = null,
+    val authorizationDetails: Collection<AuthorizationDetails>? = null,
+    val codeChallenge: String? = null,
+)