+++ CHANGELOG.md
@@ -22,7 +22,8 @@ Release 5.5.0:
    - In `WalletService`, deprecate `createCredentialRequest(CredentialRequestInput)`, provide new method `createCredentialRequest(TokenResponseParameters)` for direct processing of the token response
    - In `IssuerMetadata`, set `scope` for `SupportedCredentialFormat` to a unique string (the credential configuration id)
    - Iron out details for filtering scope and authorization details in `SimpleAuthorizationService`
-   - `SimpleAuthorizationService` correctly links credential request and issued access tokens
+   - `SimpleAuthorizationService` correctly validates requested credentials in credential request and issued access tokens
+   - `SimpleAuthorizationService` correctly validates requested credentials in authn request and token request
    - Remove proof type `cwt`, which has been removed from draft 14
  - Update dependencies:
    - Update `signum` to 3.14.0, supporting X.509 certificates in v1, v2 too